Re-organized the files and fixed CompilationEngine.c writeAndRealloc bug where errors pointer was only locally re-allocated

+++ JackAnalyzer/JackAnalyzer.c
@@ -6,10 +6,10 @@
 #include <stdlib.h>
 #include <sys/stat.h>
 
-#include "jackTokenizer.h"
+#include "../JackTokenizer/jackTokenizer.h"
 #include "JackAnalyzer.h"
 
-#include "CompilationEngine.h"
+#include "../CompilationEngine/CompilationEngine.h"
 
 void operateOnFile(char path[],char destination[]) {
     struct stat path_stat;
@@ -17,7 +17,7 @@ void operateOnFile(char path[],char destination[]) {
     const int reg = (path_stat.st_mode & _S_IFREG);
     if (reg) {
         FILE *fp = fopen(path, "rb");   // binary mode
-        if (!(int)fp) {
+        if (!fp) {
             printf("Can't open file");
             return;
         }
@@ -41,7 +41,7 @@ void operateOnFile(char path[],char destination[]) {
             const int reg_dest = S_ISREG(path_stat_dest.st_mode);
             if (reg_dest) {
                 FILE *fp_dest = fopen(destination, "wb");   // binary mode
-                if (!(int)fp_dest) return;
+                if (!fp_dest) return;
 
                 fseek(fp, 0, SEEK_END);
                 rewind(fp_dest);
